Extract mse_action_value_loss
