+++ rl_blox/algorithm/sac.py
@@ -15,10 +15,10 @@ from ..blox.function_approximator.policy_head import (
     GaussianTanhPolicy,
     StochasticPolicyBase,
 )
+from ..blox.losses import mse_action_value_loss
 from ..blox.replay_buffer import ReplayBuffer
 from ..blox.target_net import soft_target_net_update
 from ..logging.logger import LoggerBase
-from .ddpg import mse_action_value_loss
 
 
 def sac_actor_loss(
